# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "这是一个基于 Go 语言构建的企业资源规划（ERP）后端系统，使用 JWT 进行用户认证，PostgreSQL 作为数据库。API接口按功能模块分组，提供统一的路径结构。",
        "title": "ERP 后端系统 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/api/tags": {
            "get": {
                "description": "获取所有标签列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "标签管理"
                ],
                "summary": "获取所有标签",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/internal_modules_tags_model.Tag"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "创建新的标签",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "标签管理"
                ],
                "summary": "创建标签",
                "parameters": [
                    {
                        "description": "标签信息",
                        "name": "tag",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_modules_tags_model.Tag"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_modules_tags_model.Tag"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/api/tags/{id}": {
            "get": {
                "description": "根据ID获取标签详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "标签管理"
                ],
                "summary": "获取标签详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "标签ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_modules_tags_model.Tag"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "description": "更新标签信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "标签管理"
                ],
                "summary": "更新标签",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "标签ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "标签信息",
                        "name": "tag",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_modules_tags_model.Tag"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_modules_tags_model.Tag"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "description": "删除指定标签",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "标签管理"
                ],
                "summary": "删除标签",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "标签ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/tags/{id}/products": {
            "get": {
                "description": "获取指定标签下的所有产品",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "标签管理"
                ],
                "summary": "获取标签下的产品",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "标签ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.Product"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "为指定标签添加产品",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "标签管理"
                ],
                "summary": "为标签添加产品",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "标签ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "产品ID",
                        "name": "product_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "从指定标签移除产品",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "标签管理"
                ],
                "summary": "从标签移除产品",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "标签ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "产品ID",
                        "name": "product_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/oss/sts/token": {
            "get": {
                "description": "为前端直传获取阿里云OSS STS临时访问凭证",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OSS"
                ],
                "summary": "获取STS临时凭证",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/oss.STSResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_modules_product_model.Source": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "internal_modules_product_model.Tag": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_enabled": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "internal_modules_tags_model.Tag": {
            "description": "标签信息",
            "type": "object",
            "properties": {
                "color": {
                    "description": "标签颜色",
                    "type": "string",
                    "example": "#FF6B6B"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "description": {
                    "description": "标签描述",
                    "type": "string",
                    "example": "热销商品标签"
                },
                "id": {
                    "type": "integer"
                },
                "is_enabled": {
                    "description": "是否启用",
                    "type": "boolean",
                    "example": true
                },
                "name": {
                    "description": "标签名称",
                    "type": "string",
                    "example": "热销"
                },
                "products": {
                    "description": "关联的商品",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Product"
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "model.Color": {
            "description": "商品颜色信息",
            "type": "object",
            "properties": {
                "code": {
                    "description": "颜色代码",
                    "type": "string",
                    "example": "BLACK"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "hex_color": {
                    "description": "十六进制颜色值",
                    "type": "string",
                    "example": "#000000"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "description": "颜色名称",
                    "type": "string",
                    "example": "黑色"
                },
                "products": {
                    "description": "关联的商品",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Product"
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "model.Product": {
            "description": "商品信息",
            "type": "object",
            "properties": {
                "colors": {
                    "description": "颜色列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Color"
                    }
                },
                "cost_price": {
                    "description": "进货价",
                    "type": "number",
                    "example": 5999
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "discount_price": {
                    "description": "优惠价格",
                    "type": "number",
                    "example": 6799
                },
                "id": {
                    "type": "integer"
                },
                "images": {
                    "description": "商品图片列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ProductImage"
                    }
                },
                "is_discounted": {
                    "description": "是否优惠",
                    "type": "boolean",
                    "example": true
                },
                "is_enabled": {
                    "description": "是否启用",
                    "type": "boolean",
                    "example": true
                },
                "name": {
                    "description": "商品名称",
                    "type": "string",
                    "example": "iPhone 14"
                },
                "price": {
                    "description": "售价",
                    "type": "number",
                    "example": 6999
                },
                "product_code": {
                    "description": "商品编码（店铺编号-货号）",
                    "type": "string",
                    "example": "APPLE001-IPHONE14-128G-BLACK"
                },
                "shipping_time": {
                    "description": "发货时间",
                    "type": "string",
                    "example": "三天"
                },
                "sku": {
                    "description": "货号",
                    "type": "string",
                    "example": "IPHONE14-128G-BLACK"
                },
                "source": {
                    "description": "关联的货源信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/internal_modules_product_model.Source"
                        }
                    ]
                },
                "source_id": {
                    "description": "货源ID",
                    "type": "integer",
                    "example": 1
                },
                "tags": {
                    "description": "标签列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_modules_product_model.Tag"
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "model.ProductImage": {
            "description": "商品图片信息",
            "type": "object",
            "properties": {
                "alt": {
                    "description": "图片描述",
                    "type": "string",
                    "example": "iPhone 14 正面照"
                },
                "is_main": {
                    "description": "是否为主图",
                    "type": "boolean",
                    "example": true
                },
                "sort": {
                    "description": "排序，数字越小越靠前",
                    "type": "integer",
                    "example": 1
                },
                "title": {
                    "description": "图片标题",
                    "type": "string",
                    "example": "产品主图"
                },
                "url": {
                    "description": "图片URL",
                    "type": "string",
                    "example": "https://example.com/image1.jpg"
                }
            }
        },
...
        "oss.STSResponse": {
            "type": "object",
            "properties": {
                "accessKeyId": {
                    "type": "string"
                },
                "accessKeySecret": {
                    "type": "string"
                },
                "bucket": {
                    "type": "string"
                },
                "endpoint": {
                    "type": "string"
                },
                "expiration": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "securityToken": {
                    "type": "string"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "输入 \"Bearer \" 加上 JWT token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}